refactor add sections services to reduce duplicate code
