+++ app/services/JourneyAddSectionsService.scala
@@ -16,48 +16,9 @@
 
 package services
 
-import models.TagStatus.{CannotStartYet, Completed, NotStarted}
-import models.{IncomeFromWorkDependentStates, SectionState, TagStatus, UserAnswers}
-import pages.aboutyou.TaxAvoidanceSchemesPage
-import pages.pensions.PaymentsIntoPensionsPage
-import pages.propertypensionsinvestments.UkDividendsSharesLoansPage
-import pages.workandbenefits.{AboutYourWorkPage, JobseekersAllowancePage}
+import models.{SectionState, UserAnswers}
 
 class JourneyAddSectionsService extends AddSectionsService {
-  def getState(userAnswers: Option[UserAnswers]): SectionState = {
+  def getState(userAnswers: Option[UserAnswers]): SectionState = deriveState(userAnswers, isPrivateBeta = false)
 
-    userAnswers match {
-      case None => SectionState(NotStarted, CannotStartYet, CannotStartYet, NotStarted)
-      case Some(ua) =>
-        val aboutYou: TagStatus = if (ua.get(TaxAvoidanceSchemesPage).isDefined) {
-          Completed
-        } else {
-          NotStarted
-        }
-
-        val incomeFromWorkStates: IncomeFromWorkDependentStates = IncomeFromWorkDependentStates(
-          aboutYou.isCompleted,
-          ua.get(AboutYourWorkPage).isDefined,
-          ua.get(JobseekersAllowancePage).isDefined
-        )
-
-        val incomeFromProperty: TagStatus =
-          ua.get(UkDividendsSharesLoansPage) match {
-            case Some(_) => Completed
-            case None => if (incomeFromWorkStates.getStatus.isCompleted) {
-              NotStarted
-            } else {
-              CannotStartYet
-            }
-          }
-
-        val pensions: TagStatus = if (ua.get(PaymentsIntoPensionsPage).isDefined) {
-          Completed
-        } else {
-          NotStarted
-        }
-
-        SectionState(aboutYou, incomeFromWorkStates.getStatus, incomeFromProperty, pensions)
-    }
-  }
 }
